agent: remove workarounds for spatial_reuse.
Wi-Fi 6 Capability TLV are included in the AP Capability Report Message.
However, all the parameters (including wifi6_caps->spatial_reuse) within
these TLVs are set to 0.
This made it impossible for BSS coloring to work.
That's why workaround was added, which set the spatial_reuse parameter to 1.

The problem with the parameters should be fixed in PPM-3028.
Workaround is no longer needed.

Remove workarounds and add proper conditions for handling spatial_reuse.

PPM-2602.

Signed-off-by: Dmytro Puz <d.puz@inango-systems.com>