Framework: nbapi: don't stop and delete amrt objects in destructor
Ticket: PPM-3032

Description: the controller and agent process will generate segment fault due
to meta data corruption when free the memory.
Current investigation leads to the reuse of global static variables, such as,
amxc_array_types and some others. But the exact mechanism that how it corrupted
the meta data is still unclear.

Changes: skipped the amxrt object destruction, as a temporary solution. which
has limited impact to the current code base because it happens at the exit time.

Note: need further investigation to correctly destruct the amxrt object.

Signed-off-by: Lu Dai <lu.dai@mind.be>