transport: ieee1905: drain buffer before apply BPF
BPF does not affect existed frames in kernel buffer, which means,
after create socket, and before apply BPF, kernel buffer may have
queued frames, which is expected to be filtered out.
It can make the recv function, which assumes all frames it is handling
are filtered, complain. When the traffic is heavy enough, it presents
as log flood.

This change drains the buffer before apply BPF, as a result, re-open
the error log in the recv function.

Signed-off-by: Lu Dai <lu.dai@mind.be>