agent: metrics: fill AP Radio Metrics TLV
Starting with EM R2, AP Metrics Response contains one
AP Radio Metric TLV per radio.

It needs to be filled with certain information in a certain encoding.
Noise is encoded on a [0..220] scale
Transmit, ReceiveSelf and ReceiveOther contain the
percentage of time, linearly scaled with 255 representing 100%,
over the duration of the measurement.

Signed-off-by: Iacob Juc <iacob.juc_ext@softathome.com>