agent: traffic_separation: apply rules for new interface
Currently, when a newly onboarding agent associates with the BH VAP, an
interface such as wlanx.x.stax is created for each connection, Prplmesh
doesn't handle these dynamically created interfaces.

with this changes, if the agent connects to the BH VAP - the newly
created interface will added to br-lan and corresponding Traffic
Separation rules are applied.

closes PPM-3057

Signed-off-by: rramasamy <rramasamy@maxlinear.com>