bcl: ucc: port re-bind error when restart process
beerocks_agent will raise abort in certification test running in agent mode.
it is caused by re-bind to the exact same socket address when the restart
script is called. when start again, the socket is still in listen state,rather
than "time_wait" state, which makes address reusing not applicable and fail.

current TcpServerSocket/Socket/ServerSocket doesn't set the SOCK_CLOEXEC flag
by default, the fd will be inherit by child process like beerocks_fronthaul.
adding that the master branch hasn't merge the fix:
  "dev/pgrep-proc-name-length-limit"
and the beerocks_fronthaul process can not be killed by the script but assumed
to be killed. Means it is stilling holding the socket for a period when one
believe it is not.

this fix append the SOCK_CLOEXEC flag to the UccServer holding socket, choose
not to change the Socket/ServerScoket/TcpServerSocket class, which has bigger
impact (coverage test). Also, leave an info log monitoring the future code
change.

closes: ppm-3076

Signed-off-by: Lu Dai <lu.dai@mind.be>