PPM-3077 Handling 3rd party EM Controllers and EM Agents
Description :

This is on handling third party devices which are not EM+ devices.
As the EM Controller has the network master role, there shall only be one in every network.
Since there is no standardized EM Controller election or negotiation mechanism defined,
there can be conflicts when an EM network is built with devices from several device vendors.
To establish a trust relationship and determine which EM Controller and EM Agents can be trusted in the
network, this implemented method is used to handle insertion of third part EM Controllers and
EM Agents when inserted in the same network as the EM+ Controller/Agent.

Scope :

To have a config variable to enable/disable the feature to handle third party.
Added new supported service TLV parameter values such as below:

1. Multi-AP Controller: 0x00
2. Multi-AP Agent: 0x01
3. EM+ Controller: 0xA0 (New)
4. EM+ Agent: 0xA1 (New)

If the supported service TLV has any of these new values then that device will be considered as EM+ devices.
If the handling feature is enabled then only EM+ devices will be onboarded.
If the handling feature is disabled then any multi-AP devices can be onboarded.

Files Modified:
        modified:   agent/config/beerocks_agent.conf.in
        modified:   agent/src/beerocks/slave/tasks/ap_autoconfiguration_task.cpp
        modified:   agent/src/beerocks/slave/tasks/topology_task.cpp
        modified:   common/beerocks/bcl/include/bcl/beerocks_config_file.h
        modified:   common/beerocks/bcl/source/beerocks_config_file.cpp
        modified:   framework/tlvf/AutoGenerated/include/tlvf/wfa_map/tlvSupportedService.h
        modified:   framework/tlvf/yaml/tlvf/wfa_map/tlvSupportedService.yaml

Closes PPM-3077

Signed-off-by: Mahantesh <mahantesh.k@tataelxsi.co.in>
Co-Authored-by: Abijith Nayak <abijith.r@tataelxsi.co.in>