fronthaul_manager: ap_manager: failed connection message
Failed connection message should report the interface on which failed connection
is reported.

The current code adds the BSSID TLV to Failed connection message only
for MultiAP Profiles greater than or equal to 3. This condition is modified
such that the BSSID information will be available for all profiles, which can be really
helpful for diagnostic purposes.

Closes PPM-3097

Signed-off-by: abijith <abijith.r@tataelxsi.co.in>
Co-Authored-by: Subhashini.Dhanapal <subhashini.d@tataelxsi.co.in>
Signed-off-by: Mahantesh <mahantesh.k@tataelxsi.co.in>