LinkMetricsCollectionTask: fix alias usage and signal_dbm type
The function `search_path_assocDev_by_mac` was using an incorrect alias,
causing the Receiver Link Metric's RSSI to always be reported as 0.
This led to inaccurate link metric calculations.

Additionally, in AP Metric Response, `signal_dbm` was incorrectly stored as
a `uint8_t` instead of an `int8_t`, causing RCPI to always be reported as 220.

To fix these issues:
- Corrected alias usage in `search_path_assocDev_by_mac`, ensuring proper RSSI reporting.
- Updated `signal_dbm` type to `int8_t` to prevent incorrect RCPI values.

Issue: PPW-396

Signed-off-by: Nefi Guclu <nefi.guclu@airties.com>