agent: handle_link_metric_query: Ignore AL MAC for ALL_NEIGHBORS query
This commit addresses a interpretation difference related to section 6.4.10
(Link Metric Query TLV) of the IEEE 1905.1-2013 document.

In the same document, in Table 6-16 (Link Metric Query), tlvLength is 8, and
the following statement regarding the first octet of tlvValue is seen:
- If the value is 0 (All neighbors), then the EUI-48 field is not present;
- If the value is 1 (Specific neighbor), then the EUI-48 field shall be present.

Two different interpretations arise from this statement:
- Based on the phrase "not present," the EUI-48 field should be omitted from
the TLV in the All Neighbors case. Therefore, tlvLength will be 2.
- Since tlvLength is specified as 8 in the document, the EUI-48 field should
always be included in the TLV. Therefore, tlvLength will be 8.

It can be seen that these two statements cause confusion in some IEEE 1905
implementations. Check out the open-source IEEE 1905 implementation at the
link below. It explains how the two situations mentioned above are handled:
https://github.com/BroadbandForum/meshComms/blob/master/src/factory/src_independent/1905_tlvs.c#L2125

In short, even in the All Neighbors case, tlvLength is always 8. The EUI-48
field is not removed from the TLV but is instead zeroed out, and the first
byte is set to the "link metrics requested" value.
This ensures compatibility with implementations like the prplMesh agent that
accept cases where tlvLength is 2. In other words, since the "link metrics
requested" parameter is shifted to the first byte of the MAC address, the
desired data will still be accessible.

To clarify the issue more clearly: The AirTies controller takes the above-mentioned
scenario into account when generating the Link Metric Query. That is, the TLV
length is always 8, and in the "All neighbors" case, the "link metrics requested"
parameter is shifted to the first byte of the MAC address.
In the current prplMesh agent, there are two different Link Metric Query TLV
classes. One of them, tlvLinkMetricQuery(tlvLength: 8), includes the EUI-48
field, while the other, tlvLinkMetricQueryAllNeighbors(tlvLength: 2), does not.

When the prplMesh agent receives a Link Metric Query with a tlvLength of 2, it
converts it to the tlvLinkMetricQueryAllNeighbors class, and a query is made for
the All neighbors case.

When the prplMesh agent receives a Link Metric Query with a tlvLength of 8, it
is converted to the tlvLinkMetricQuery class, and a query is made for the MAC
address within the TLV. In tlvLinkMetricQuery, when the query type is "All neighbors,"
the EUI-48 field is irrelevant and must be ignored. However, the prplMesh agent
does not ignore this field; instead, it performs a query based on the MAC address
in that field. In the scenario where the "link metrics requested" parameter is
copied to the first byte of the MAC address field, a non-zero MAC address will
be encountered. This results in the query being made as if it's for a specific
neighbor, even in the All neighbors case.

To avoid this situation, an improvement was made in the prplMesh agent based on
the link metric query type. In the All neighbors case, it is ensured that the
MAC address will always be ZERO_MAC.

PPM-3192

Signed-off-by: Durmus Koyuncu <durmus.koyuncu@airties.com>