nbapi: ambiorix: connect to all the bus
Currently the connection is only made with a single fixed bus provided
(ubus by default), we want to connect to all the bus possible to be
able to access the datamodel on every bus.

We can adapt the libamxrt to not use its own event loop and connect to
all the bus detected.
However we need to be carful, because there is already a connection
done before by an other part of the code, a reset of the config may
be needed if adding more connection, or changing the way it work.

Remove already initialized variables, change the config to not auto
connect to the bus, use libamxrt to detect all the backends, load
them, and fill the config with there info, connect and register the
datamodel to all the bus, and register them to the event loop.

PPW-449

Signed-off-by: Thomas Roussarie <thomas.roussarie@softathome.com>