nbapi: ambiorix: update test for bus connection
Prepare for "nbapi: ambiorix: connect to all the bus".

Change tests to be able to work with the new connection to all bus.

Add a config as we doesn't use backend and uri as string anymore,
add amxrt class as it is used to initialize amxrt struct, and clean
it.

PPW-449

Signed-off-by: Thomas Roussarie <thomas.roussarie@softathome.com>