agent: son_slave_thread: Dynamically detect LAN interfaces
The prplmesh agent reads LAN interfaces from uci during initialization.
If the LAN interface list is not present in the uci configuration,
the default interface list is used.
Since device-to-device LAN interface names can vary, it is not feasible
to maintain all possible names in a default list. So, setting LAN
interfaces in the uci config is critically important.
To eliminate all these concerns, the idea of detecting LAN interfaces
at runtime emerges.

The core idea is to obtain LAN interfaces by filtering out WLAN and
non-physical interfaces from bridge members.

To summarize the changes;
- All interfaces connected to any bridge are evaluated to determine
  whether they are LAN interfaces.
- WLAN and non-physical interfaces are skipped/excluded.
- Removed unused function "beerocks::bpl::bpl_get_lan_interfaces()"

PPM-3245

Signed-off-by: Durmus Koyuncu <durmus.koyuncu@airties.com>