agent: ap_autoconfiguration_task: agent sends wrong center frequency
During Easymesh onboarding, in SLAVE_JOINED_NOTIFICATION the center frequency
for 6GHz with 160MHz was sent wrong.
Instead of sending center frequency, Agent is sending primary_80MHz center
frequency for 6GHz in 160MHz.

It should be m_center_frequency_2 and not m_center_frequency for 6GHz if the
BW is 160MHz.

Closes PPM-3325

Signed-off-by: rramasamy <rramasamy@maxlinear.com>